Update API and add command line items
Update API with below fields
resourceName, resourceNamespace:
get the resource name and namespace from the executed action to match
the yaml if the path is a directory. If name is missing and path is a directory
use the changeRequest name. If namespace is missing use the
changeRequest namespace.

The following flags are added
git-secret-namespace, git-secret-name, git-username, git-email